+++ odoo/songs/upgrade/10_27_1_update.py
@@ -26,21 +26,6 @@ def update_translation_for_specific_followup(ctx):
 
 
 @anthem.log
-def remove_customer_supplier_balance(ctx):
-    """ Remove customer and supplier balance """
-    balance_customer = ctx.env.ref(
-        '__setup__.account_move_balance_customer', raise_if_not_found=False)
-    if balance_customer:
-        balance_customer.unlink()
-
-    balance_supplier = ctx.env.ref(
-        '__setup__.account_move_balance_supplier', raise_if_not_found=False)
-    if balance_supplier:
-        balance_supplier.unlink()
-
-
-@anthem.log
 def post(ctx):
     """ POST 10.27.1 """
     update_translation_for_specific_followup(ctx)
-    remove_customer_supplier_balance(ctx)