+++ HISTORY.rst
@@ -9,6 +9,7 @@ Unreleased
 **Features and Improvements**
 
 * ALCYN-1504: Display the intrastat table inside the main table (with products) to avoid break page for small invoice
+* ALCYN-1576: ALCYN-1576: Remove customer/supplier balances import. Balances will be imported later (after go live)
 
 **Bugfixes**
 
@@ -38,7 +39,6 @@ Unreleased
 * ALCYN-1228: Hide the field "US Code" in the view product.template. This field duplicate the field Intrastat
 * ALCYN-459: Add a warning on sale order lines for a product 'Acheté-Vendu'
 * ALCYN-1516: Change default language of new partners to "French (BE)"
-* ALCYN-1576: ALCYN-1576: Remove customer/supplier balances import. Balances will be imported later (after go live)
 
 **Bugfixes**